create controller class, add item route

+++ Entities/Item.cs
@@ -1,11 +1,10 @@
-namespace Scith.Entities
-{
+namespace Scith.Entities{
     public record Item{
 
         //only allow setting values during init
         public Guid Id {get; init;} 
         public string Name {get; init;}
-        public decimal Price {get; init;}
+        public double Price {get; init;}
         public DateTimeOffset CreatedDate {get; init;}
     }
 }
